refactor(core): split NbtPathModel and NbtPredicate for SRP compliance
- NbtPathModel.kt now contains only path AST types (NbtPathNode, NbtPathRepr)
- NbtPredicateModel.kt holds predicate model types (NbtCompoundPredicate, etc.)
- NbtPredicateScope.kt holds the public predicate DSL surface
- NbtPredicateBuilder.kt holds the internal builder implementation
- Remove scope reference from NbtPredicateKey (leaked mutable state)
- Fix BlockNbt.kt KDoc example to use structured API
- Add typed matcher factory overloads for NbtPath and EntitySelector
